♻️ : 이벤트 관련 유틸리티 분리
