+++ src/js/display/YoutubeClassRoom/SearchForm.js
@@ -1,7 +1,8 @@
 import { $ } from '@Utils/Dom';
 import { SELECTOR } from '@Constants/Selector';
 import { ERROR_MESSAGE, ACTION_TYPE } from '@Constants/String';
-import { onEnableButton, addEventDelegate } from '@Utils/ElementControl';
+import { onEnableButton } from '@Utils/ElementControl';
+import { addEventDelegate } from '@Utils/CustomEvent';
 import YoutubeSearchStore from '@Domain/YoutubeSearchStore';
 import { isEmptyString, isSameKeyword } from '@Utils/Validator';
