+++ src/js/utils/ElementControl.js
@@ -21,18 +21,3 @@ export const onObserveElement = ($element, handler) => {
   );
   scrollObserver.observe($element);
 };
-
-export const addEventDelegate = (
-  container,
-  selector,
-  { eventType, handler, defaultEvent = false },
-) => {
-  const children = [...container.querySelectorAll(selector)];
-  const isTarget = target => children.includes(target) || target.closest(selector);
-
-  container.addEventListener(eventType, event => {
-    if (defaultEvent === true) event.preventDefault();
-    if (!isTarget(event.target)) return false;
-    handler(event);
-  });
-};
